Moved remaining ITs out of elastiknn-testing
