+++ elastiknn-plugin/src/it/resources/com/klibisz/elastiknn/testdata-sparsebool.json.gz
