+++ elastiknn-plugin/src/it/resources/log4j.properties
@@ -1,0 +1,4 @@
+log4j.rootLogger=INFO, out
+log4j.appender.out=org.apache.log4j.ConsoleAppender
+log4j.appender.out.layout=org.apache.log4j.PatternLayout
+log4j.appender.out.layout.conversionPattern=[%d{ISO8601}][%-5p][%-25c] %m%n
