+++ elastiknn-plugin/src/it/scala/com/klibisz/elastiknn/FunctionScoreQuerySuite.scala
@@ -1,8 +1,7 @@
-package com.klibisz.elastiknn.query
+package com.klibisz.elastiknn
 
 import com.klibisz.elastiknn.api._
 import com.klibisz.elastiknn.models.ExactModel
-import com.klibisz.elastiknn.testing.ElasticAsyncClient
 import com.sksamuel.elastic4s.ElasticDsl._
 import futil.Futil
 import org.scalatest._
