+++ elastiknn-plugin/src/it/scala/com/klibisz/elastiknn/QueryRescorerSuite.scala
@@ -1,7 +1,6 @@
-package com.klibisz.elastiknn.query
+package com.klibisz.elastiknn
 
 import com.klibisz.elastiknn.api._
-import com.klibisz.elastiknn.testing.ElasticAsyncClient
 import com.sksamuel.elastic4s.ElasticDsl._
 import futil.Futil
 import org.scalatest._
